wayland: don't ignore the mouse's X coordinate when selecting
Before this patch, the mouse cursor's X coordinate was ignored when
updating the currently selected item. This meant that you could
change (and activate) items while the cursor was inside the left or
right margins.

It also fixes an issue where clicking *outside* the window could
activate an item. This was only possible on compositors that allows
sending mouse events to a client, even when the mouse is outside the
window (and no buttons being held/dragged). Sounds buggy to me, but
who am I to judge 🤷?

Closes #606