Rework --accept-nth and --with-nth to accept a format string
With this, --accept-nth and --with-nth can either be specified as a
single integer (--with-nth=2), or as a template
string (--with-nth='{1}').

The simplified case, --with-nth=1, is converted to '{1}' internally -
expansion always assumes a template string.

If the parameter is in any way invalid, it is copied verbatim instead
of being expanded. This includes:

* Non-closed '{'
* '{x}', where x is not an integer
* '{x}', where x is an integer, but larger than the number of
  available column

Not planned:

* fzf-styled: simplified variants: --with-nth=1,2,
  --with-nth=1..2 (you'll need to use a template string instead)

Closes #598