meson: add fish-based test scripts
This allows you to execute the tests by simply running "ninja test" or
"meson test".

The path to the fuzzel binary to use is passed via the environment
variable $FUZZEL_TEST_BIN.

The tests are *not* allowed to run in parallel, since fuzzel exits
with an error if there's already another fuzzel instance running.