main: regression: --{with,accept}-nth=0 should be "ignored" (#671)
As documented in the man page, `--with-nth=0` and `--accept-nth=0` is the same as not specifying them at all.

This regressed when we added support for custom string formats.

Note: this is not the same as `--with-nth={0}`, which is just an invalid formatter, that **will** result in `"{0}"` in the output string.

Closes #670

Co-authored-by: Mark Stosberg <mark@stosberg.com>
Reviewed-on: https://codeberg.org/dnkl/fuzzel/pulls/671
Co-authored-by: Daniel Eklöf <daniel@ekloef.se>
Co-committed-by: Daniel Eklöf <daniel@ekloef.se>