dmenu: add --only-match
By default, fuzzel allows the user to execute/select text that isn't
a listed entry.

In application mode, this allows the user to type custom command
lines.

In dmenu mode, it allows the user to input/select text not in the
list.

Sometimes it's desirable to prevent this in dmenu mode, forcing the
user to either select one of the available entries, or abort. This is
what --only-match does.

Closes #651