config: add --mesg, --mesg-mode and --message-color command line options
and the corresponding config options:

* mesg
* mesg-mode
* colors.message

--mesg can be used to display an informational message above the
prompt. The default is to wrap long messages, but this can be changed
to instead automatically expand the window with
--mesg-mode=wrap|expand.

The color of the message text can be changed with --message-color.

This feature is likely most useful in dmenu mode, but it isn't limited
to it. I.e. it can be used in application mode too.

Closes #454